feat: add native TTS service integration for audiobook playback
